django-apps/placed_backend/project_platform/serializers.py
from rest_framework import serializers
from project_platform.models import ProjectPlatform


class ProjectPlatformSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProjectPlatform
