django-apps/placed_backend/requirements.txt
Django==1.8.3
django-cors-headers==1.1.0
django-filter==0.11.0
django-mysql==0.2.2
djangorestframework==3.2.2
Markdown==2.6.2
mysql==0.0.1
MySQL-python==1.2.5
Pillow==3.0.0
simplejson==3.8.0
